+++ src/main/java/com/buuz135/industrial/tile/block/FluidPumpBlock.java
@@ -1,15 +1,21 @@
 package com.buuz135.industrial.tile.block;
 
 import com.buuz135.industrial.book.BookCategory;
+import com.buuz135.industrial.config.CustomConfiguration;
 import com.buuz135.industrial.proxy.ItemRegistry;
 import com.buuz135.industrial.tile.world.FluidPumpTile;
 import com.buuz135.industrial.utils.RecipeUtils;
+import lombok.Getter;
 import net.minecraft.block.material.Material;
 import net.minecraft.init.Items;
 import net.minecraft.item.ItemStack;
+import net.minecraftforge.common.config.Configuration;
 import net.ndrei.teslacorelib.items.MachineCaseItem;
 
 public class FluidPumpBlock extends CustomAreaOrientedBlock<FluidPumpTile> {
+
+    @Getter
+    private boolean replaceFluidWithCobble;
 
     public FluidPumpBlock() {
         super("fluid_pump", FluidPumpTile.class, Material.ROCK, 1000, 40, RangeType.UP, 1, 1, true);
@@ -24,6 +30,12 @@ public class FluidPumpBlock extends CustomAreaOrientedBlock<FluidPumpTile> {
                 'w', Items.WATER_BUCKET,
                 'm', MachineCaseItem.INSTANCE,
                 'g', "gearGold");
+    }
+
+    @Override
+    public void getMachineConfig() {
+        super.getMachineConfig();
+        replaceFluidWithCobble = CustomConfiguration.config.getBoolean("replaceFluidWithCobble", "machines" + Configuration.CATEGORY_SPLITTER + this.getRegistryName().getResourcePath().toString(), true, "Fluid pump should replace the picked up fluid with cobblestone");
     }
 
     @Override